Move raw_seconds_short to beets.util
